Utilizing debounce
