+++ packages/usa-character-count/src/index.js
@@ -1,5 +1,6 @@
 const select = require("../../uswds-core/src/js/utils/select");
 const behavior = require("../../uswds-core/src/js/utils/behavior");
+const debounce = require("../../uswds-core/src/js/utils/debounce");
 const { prefix: PREFIX } = require("../../uswds-core/src/js/config");
 
 const CHARACTER_COUNT_CLASS = `${PREFIX}-character-count`;
@@ -142,35 +143,15 @@ const updateSrStatus = (msgEl, statusMessage, assertive = false) => {
 };
 
 /**
- * @param {Function} callback
- * @param {number} delay
- * @returns {Function & { cancel: () => void }}
- */
-const createDebounced = (callback, delay) => {
-  let timer = null;
-  const debounced = (...args) => {
-    window.clearTimeout(timer);
-    timer = window.setTimeout(() => {
-      callback(...args);
-    }, delay);
-  };
-  debounced.cancel = () => {
-    window.clearTimeout(timer);
-    timer = null;
-  };
-  return debounced;
-};
-
-/**
  * Updates the character count status for screen readers after typing pauses.
  *
  * @param {HTMLElement} msgEl - The screen reader status message element
  * @param {string} statusMessage - A string of the current character status
  * @param {boolean} assertive - Whether to announce assertively when over the limit
  */
-const srUpdateStatus = createDebounced(updateSrStatus, SR_STATUS_DEBOUNCE_MS);
+const srUpdateStatus = debounce(updateSrStatus, SR_STATUS_DEBOUNCE_MS);
 
-const srUpdateStatusRecovery = createDebounced(updateSrStatus, AT_DEFER_MS);
+const srUpdateStatusRecovery = debounce(updateSrStatus, AT_DEFER_MS);
 
 const validitySyncTimers = new WeakMap();
